feat: move comment, location and comment into values [sc-69]
