+++ packages/server/invites/src/domain/entities/__init__.py
@@ -1,5 +1,2 @@
 from .user_auth import UserAuth
-from .comment import Comment
 from .invite import Invite
-from .location import Location
-from .rating import Rating
