+++ packages/server/invites/src/domain/entities/values/rating.py
